refactor(web): consume runtime config in auth and dashboard flows
