+++ surfsense_web/app/desktop/login/layout.tsx
@@ -1,0 +1,5 @@
+import { RuntimeConfig } from "@/components/providers/runtime-config.server";
+
+export default function DesktopLoginLayout({ children }: { children: React.ReactNode }) {
+	return <RuntimeConfig>{children}</RuntimeConfig>;
+}
